Enhanced architecture: moved Constants file to utils package.
