+++ app/src/main/java/com/thewhitewings/pouch/ui/NoteViewModel.java
@@ -8,7 +8,7 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.thewhitewings.pouch.Constants;
+import com.thewhitewings.pouch.utils.Constants;
 import com.thewhitewings.pouch.data.Note;
 import com.thewhitewings.pouch.data.NotesRepository;
